refactor: reorganize seed analysis and serialization into separate modules #9 #8
- Move seed analysis functions to ncorr/algo/seed_analysis.hpp
- Extract binary I/O utilities to ncorr/io/binary_io.hpp
- Add core header wrappers under ncorr/core/ directory
- Split ncorr.cpp into ncorr_seed.cpp, ncorr_serialization.cpp, and ncorr_visualization.cpp
- Update CMakeLists.txt to include new source files
- Replace manual binary I/O with templated read_pod/write_pod helpers
- Rename SeedConvergence to SeedQuality and S
