+++ .gitignore
@@ -10,3 +10,5 @@ test/examples/.DS_Store
 test/examples/ohtcfrp/.DS_Store
 .cache
 compile_commands.json
+build-test
+bin/
