+++ include/ncorr.h
@@ -364,78 +364,6 @@ struct SeedParams final {
     }
 };
 
-struct SeedConvergence final {
-    int num_iterations;
-    double diffnorm;
-    
-    SeedConvergence() : num_iterations(0), diffnorm(0.0) { }
-    SeedConvergence(int iter, double norm) : num_iterations(iter), diffnorm(norm) { }
-};
-
-struct SeedAnalysisResult final {
-    std::vector<SeedParams> seeds;
-    std::vector<SeedConvergence> convergence;
-    bool success;
-    
-    SeedAnalysisResult() : success(false) { }
-};
-
-// Lightweight seed analysis for predicting reference updates
-SeedAnalysisResult analyze_seeds(
-    const details::subregion_nloptimizer &sr_nloptimizer,
-    const Array2D<double>& ref_gs,
-    const ROI2D& roi,
-    const std::vector<SeedParams>& seed_positions,
-    ROI2D::difference_type radius,
-    int cutoff_iteration,
-    double cutoff_max_diffnorm,
-    double cutoff_max_corrcoef,
-    bool debug = true
-);
-
-// Structure to hold precomputed seed data for each frame
-struct SeedComputationData final {
-    ROI2D roi;                    // Updated ROI for this frame
-    std::vector<SeedParams> seed_params_by_region;  // One seed param per region
-    details::subregion_nloptimizer sr_nloptimizer;
-    
-    SeedComputationData() = default;
-    SeedComputationData(const ROI2D& roi, const std::vector<SeedParams>& params, const details::subregion_nloptimizer &sr_nloptimizer) : 
-        roi(roi), seed_params_by_region(params), sr_nloptimizer(sr_nloptimizer) { }
-};
-
-// Compute seed parameters for all frames with ROI updates
-// Returns precomputed data (roi, seed_params) for each successfully analyzed frame
-std::vector<SeedComputationData> compute_only_seed_points(
-    const Array2D<double>& A_ref,
-    const std::vector<Array2D<double>>& A_curs,
-    const ROI2D& roi,
-    ROI2D::difference_type scalefactor,
-    INTERP interp_type,
-    SUBREGION subregion_type,
-    const std::vector<SeedParams>& seeds_by_region,  // One seed per region
-    double cutoff_corrcoef,
-    ROI2D::difference_type region_idx = 0,
-    bool debug = false
-);
-
-// Compute displacements using precomputed seed parameters
-Disp2D compute_displacements(
-    const details::subregion_nloptimizer &sr_nloptimizer,
-    const ROI2D& roi_reduced,
-    const SeedParams& seedparams,
-    ROI2D::difference_type scalefactor,
-    double cutoff_corrcoef,
-    ROI2D::difference_type region_idx,
-    bool debug
-) ;
-
-// Update seed positions based on displacement (seed propagation)
-std::vector<SeedParams> propagate_seeds(
-    const std::vector<SeedParams>& seeds,
-    ROI2D::difference_type spacing
-);
-
 struct DIC_analysis_parallel_input final {
     typedef ROI2D::difference_type                              difference_type;
