+++ include/ncorr/core/array2d.hpp
@@ -1,0 +1,6 @@
+#ifndef NCORR_CORE_ARRAY2D_HPP
+#define NCORR_CORE_ARRAY2D_HPP
+
+#include "../../Array2D.h"
+
+#endif
