+++ include/ncorr/core/data2d.hpp
@@ -1,0 +1,6 @@
+#ifndef NCORR_CORE_DATA2D_HPP
+#define NCORR_CORE_DATA2D_HPP
+
+#include "../../Data2D.h"
+
+#endif
