+++ include/ncorr/core/disp2d.hpp
@@ -1,0 +1,6 @@
+#ifndef NCORR_CORE_DISP2D_HPP
+#define NCORR_CORE_DISP2D_HPP
+
+#include "../../Disp2D.h"
+
+#endif
