+++ include/ncorr/core/image2d.hpp
@@ -1,0 +1,6 @@
+#ifndef NCORR_CORE_IMAGE2D_HPP
+#define NCORR_CORE_IMAGE2D_HPP
+
+#include "../../Image2D.h"
+
+#endif
