+++ include/ncorr/core/roi2d.hpp
@@ -1,0 +1,6 @@
+#ifndef NCORR_CORE_ROI2D_HPP
+#define NCORR_CORE_ROI2D_HPP
+
+#include "../../ROI2D.h"
+
+#endif
