+++ include/ncorr/core/strain2d.hpp
@@ -1,0 +1,6 @@
+#ifndef NCORR_CORE_STRAIN2D_HPP
+#define NCORR_CORE_STRAIN2D_HPP
+
+#include "../../Strain2D.h"
+
+#endif
