+++ include/ncorr/ncorr.hpp
@@ -1,0 +1,6 @@
+#ifndef NCORR_UMBRELLA_HPP
+#define NCORR_UMBRELLA_HPP
+
+#include "../ncorr.h"
+
+#endif
