+++ src/ncorr.cpp
@@ -2405,289 +2405,6 @@ DIC_analysis_input::DIC_analysis_input(const std::vector<Image2D> &imgs,
     }
 }  
 
-DIC_analysis_input DIC_analysis_input::load(std::ifstream &is) {
-    // Form empty DIC_analysis_input then fill in values in accordance to how they are saved
-    DIC_analysis_input DIC_input;
-    
-    // Load Images
-    difference_type num_images = 0;
-    is.read(reinterpret_cast<char*>(&num_images), std::streamsize(sizeof(difference_type)));
-    DIC_input.imgs.resize(num_images);
-    for (auto &img : DIC_input.imgs) {
-        img = Image2D::load(is);
-    }
-    
-    // Load roi
-    DIC_input.roi = ROI2D::load(is);
-    
-    // Load scalefactor
-    is.read(reinterpret_cast<char*>(&DIC_input.scalefactor), std::streamsize(sizeof(difference_type)));
-    
-    // Load interp_type
-    is.read(reinterpret_cast<char*>(&DIC_input.interp_type), std::streamsize(sizeof(INTERP)));
-    
-    // Load subregion_type
-    is.read(reinterpret_cast<char*>(&DIC_input.subregion_type), std::streamsize(sizeof(SUBREGION)));
-    
-    // Load radius
-    is.read(reinterpret_cast<char*>(&DIC_input.r), std::streamsize(sizeof(difference_type)));
-    
-    // Load number of threads
-    is.read(reinterpret_cast<char*>(&DIC_input.num_threads), std::streamsize(sizeof(difference_type)));
-    
-    // Load cutoff_corrcoef
-    is.read(reinterpret_cast<char*>(&DIC_input.cutoff_corrcoef), std::streamsize(sizeof(double)));
-    
-    // Load update_corrcoef
-    is.read(reinterpret_cast<char*>(&DIC_input.update_corrcoef), std::streamsize(sizeof(double)));
-    
-    // Load prctile_corrcoef
-    is.read(reinterpret_cast<char*>(&DIC_input.prctile_corrcoef), std::streamsize(sizeof(double)));
-    
-    // Load roi_update_mode 
-    is.read(reinterpret_cast<char*>(&DIC_input.roi_update_mode), std::streamsize(sizeof(ROI_UPDATE_MODE)));
-    
-    // Load accumulation_mode 
-    is.read(reinterpret_cast<char*>(&DIC_input.accumulation_mode), std::streamsize(sizeof(ACCUMULATION_MODE)));
-    
-    // Load save_disps_steps
-    is.read(reinterpret_cast<char*>(&DIC_input.save_disps_steps), std::streamsize(sizeof(bool)));
-    
-    // Load debug
-    is.read(reinterpret_cast<char*>(&DIC_input.debug), std::streamsize(sizeof(bool)));
-    
-    return DIC_input;
-}
-
-DIC_analysis_input DIC_analysis_input::load(const std::string &filename) {
-    // Form stream
-    std::ifstream is(filename.c_str(), std::ios::in | std::ios::binary);
-    if (!is.is_open()) {
-        throw std::invalid_argument("Could not open " + filename + " for loading DIC_analysis_input.");
-    }
-    
-    // Form DIC_analysis_input using stream static factory method
-    auto DIC_input = DIC_analysis_input::load(is);
-    
-    // Close stream
-    is.close();
-    
-    return DIC_input;
-}
-    
-void save(const DIC_analysis_input &DIC_input, std::ofstream &os) {    
-    typedef ROI2D::difference_type                              difference_type;
-    
-    // Save imgs -> roi -> scalefactor -> interp_type -> subregion_type -> 
-    // radius -> num_threads -> cutoff_corrcoef -> update_corrcoef -> 
-    // prctile_corrcoef -> debug
-        
-    difference_type num_imgs = DIC_input.imgs.size();
-    os.write(reinterpret_cast<const char*>(&num_imgs), std::streamsize(sizeof(difference_type)));    
-    for (const auto &img : DIC_input.imgs) {
-        save(img, os);
-    }
-    
-    save(DIC_input.roi, os);
-    
-    os.write(reinterpret_cast<const char*>(&DIC_input.scalefactor), std::streamsize(sizeof(difference_type)));   
-    
-    os.write(reinterpret_cast<const char*>(&DIC_input.interp_type), std::streamsize(sizeof(INTERP)));   
-    
-    os.write(reinterpret_cast<const char*>(&DIC_input.subregion_type), std::streamsize(sizeof(SUBREGION)));   
-    
-    os.write(reinterpret_cast<const char*>(&DIC_input.r), std::streamsize(sizeof(difference_type)));   
-    
-    os.write(reinterpret_cast<const char*>(&DIC_input.num_threads), std::streamsize(sizeof(difference_type)));   
-    
-    os.write(reinterpret_cast<const char*>(&DIC_input.cutoff_corrcoef), std::streamsize(sizeof(double)));      
-    
-    os.write(reinterpret_cast<const char*>(&DIC_input.update_corrcoef), std::streamsize(sizeof(double)));   
-    
-    os.write(reinterpret_cast<const char*>(&DIC_input.prctile_corrcoef), std::streamsize(sizeof(double)));   
-    
-    os.write(reinterpret_cast<const char*>(&DIC_input.roi_update_mode), std::streamsize(sizeof(ROI_UPDATE_MODE)));   
-    
-    os.write(reinterpret_cast<const char*>(&DIC_input.accumulation_mode), std::streamsize(sizeof(ACCUMULATION_MODE)));   
-    
-    os.write(reinterpret_cast<const char*>(&DIC_input.save_disps_steps), std::streamsize(sizeof(bool)));   
-    
-    os.write(reinterpret_cast<const char*>(&DIC_input.debug), std::streamsize(sizeof(bool)));  
-}
-
-void save(const DIC_analysis_input &DIC_input, const std::string &filename) {
-    // Form stream
-    std::ofstream os(filename.c_str(), std::ios::out | std::ios::binary | std::ios::trunc);
-    if (!os.is_open()) {
-        throw std::invalid_argument("Could not open " + filename + " for saving DIC_analysis_input.");
-    }
-    
-    // Save DIC_input into stream
-    save(DIC_input, os);
-        
-    // Close stream
-    os.close();
-}
-
-DIC_analysis_output DIC_analysis_output::load(std::ifstream &is) {
-    // Form empty DIC_analysis_output then fill in values in accordance to how they are saved
-    DIC_analysis_output DIC_output;
-    
-    // Load disps
-    difference_type num_disps = 0;
-    is.read(reinterpret_cast<char*>(&num_disps), std::streamsize(sizeof(difference_type)));
-    DIC_output.disps.resize(num_disps);
-    for (auto &disp : DIC_output.disps) {
-        disp = Disp2D::load(is);
-    }
-    
-    // Load perspective type
-    is.read(reinterpret_cast<char*>(&DIC_output.perspective_type), std::streamsize(sizeof(PERSPECTIVE)));
-    
-    // Load units
-    difference_type length = 0;
-    is.read(reinterpret_cast<char*>(&length), std::streamsize(sizeof(difference_type)));
-    DIC_output.units = std::string(length,' ');
-    is.read(const_cast<char*>(DIC_output.units.c_str()), std::streamsize(length));
-        
-    // Load units per pixel
-    is.read(reinterpret_cast<char*>(&DIC_output.units_per_pixel), std::streamsize(sizeof(double)));
-    
-    return DIC_output;
-}
-
-DIC_analysis_output DIC_analysis_output::load(const std::string &filename) {
-    // Form stream
-    std::ifstream is(filename.c_str(), std::ios::in | std::ios::binary);
-    if (!is.is_open()) {
-        throw std::invalid_argument("Could not open " + filename + " for loading DIC_analysis_output.");
-    }
-    
-    // Form DIC_analysis_output using stream static factory method
-    auto DIC_output = DIC_analysis_output::load(is);
-    
-    // Close stream
-    is.close();
-    
-    return DIC_output;
-}
-    
-void save(const DIC_analysis_output &DIC_output, std::ofstream &os) {    
-    typedef ROI2D::difference_type                              difference_type;
-    
-    // Save disps
-    difference_type num_disps = DIC_output.disps.size();
-    os.write(reinterpret_cast<const char*>(&num_disps), std::streamsize(sizeof(difference_type)));    
-    for (const auto &disp : DIC_output.disps) {
-        save(disp, os);
-    }    
-    
-    // Save perspective type
-    os.write(reinterpret_cast<const char*>(&DIC_output.perspective_type), std::streamsize(sizeof(PERSPECTIVE)));   
-        
-    // Save units
-    difference_type length = DIC_output.units.size();
-    os.write(reinterpret_cast<const char*>(&length), std::streamsize(sizeof(difference_type)));
-    os.write(DIC_output.units.c_str(), std::streamsize(length));
-    
-    // Save units per pixel
-    os.write(reinterpret_cast<const char*>(&DIC_output.units_per_pixel), std::streamsize(sizeof(double))); 
-}
-
-void save(const DIC_analysis_output &DIC_output, const std::string &filename) {
-    // Form stream
-    std::ofstream os(filename.c_str(), std::ios::out | std::ios::binary | std::ios::trunc);
-    if (!os.is_open()) {
-        throw std::invalid_argument("Could not open " + filename + " for saving DIC_analysis_output.");
-    }
-    
-    // Save DIC_output into stream
-    save(DIC_output, os);
-        
-    // Close stream
-    os.close();
-}
-
-// DIC_analysis_step_data save/load -------------------------------------------//
-DIC_analysis_step_data DIC_analysis_step_data::load(std::ifstream &is) {
-    typedef ROI2D::difference_type                              difference_type;
-    
-    DIC_analysis_step_data step_data;
-    
-    // Load step_disps
-    difference_type num_disps = 0;
-    is.read(reinterpret_cast<char*>(&num_disps), std::streamsize(sizeof(difference_type)));
-    step_data.step_disps.resize(num_disps);
-    for (auto &disp : step_data.step_disps) {
-        disp = Disp2D::load(is);
-    }
-    
-    // Load step_rois
-    difference_type num_rois = 0;
-    is.read(reinterpret_cast<char*>(&num_rois), std::streamsize(sizeof(difference_type)));
-    step_data.step_rois.resize(num_rois);
-    for (auto &roi : step_data.step_rois) {
-        roi = ROI2D::load(is);
-    }
-    
-    // Load step_ref_idx
-    difference_type num_ref_idx = 0;
-    is.read(reinterpret_cast<char*>(&num_ref_idx), std::streamsize(sizeof(difference_type)));
-    step_data.step_ref_idx.resize(num_ref_idx);
-    for (auto &idx : step_data.step_ref_idx) {
-        is.read(reinterpret_cast<char*>(&idx), std::streamsize(sizeof(difference_type)));
-    }
-    
-    return step_data;
-}
-
-DIC_analysis_step_data DIC_analysis_step_data::load(const std::string &filename) {
-    std::ifstream is(filename.c_str(), std::ios::in | std::ios::binary);
-    if (!is.is_open()) {
-        throw std::invalid_argument("Could not open " + filename + " for loading DIC_analysis_step_data.");
-    }
-    
-    auto step_data = DIC_analysis_step_data::load(is);
-    is.close();
-    
-    return step_data;
-}
-
-void save(const DIC_analysis_step_data &step_data, std::ofstream &os) {
-    typedef ROI2D::difference_type                              difference_type;
-    
-    // Save step_disps
-    difference_type num_disps = step_data.step_disps.size();
-    os.write(reinterpret_cast<const char*>(&num_disps), std::streamsize(sizeof(difference_type)));
-    for (const auto &disp : step_data.step_disps) {
-        save(disp, os);
-    }
-    
-    // Save step_rois
-    difference_type num_rois = step_data.step_rois.size();
-    os.write(reinterpret_cast<const char*>(&num_rois), std::streamsize(sizeof(difference_type)));
-    for (const auto &roi : step_data.step_rois) {
-        save(roi, os);
-    }
-    
-    // Save step_ref_idx
-    difference_type num_ref_idx = step_data.step_ref_idx.size();
-    os.write(reinterpret_cast<const char*>(&num_ref_idx), std::streamsize(sizeof(difference_type)));
-    for (const auto &idx : step_data.step_ref_idx) {
-        os.write(reinterpret_cast<const char*>(&idx), std::streamsize(sizeof(difference_type)));
-    }
-}
-
-void save(const DIC_analysis_step_data &step_data, const std::string &filename) {
-    std::ofstream os(filename.c_str(), std::ios::out | std::ios::binary | std::ios::trunc);
-    if (!os.is_open()) {
-        throw std::invalid_argument("Could not open " + filename + " for saving DIC_analysis_step_data.");
-    }
-    
-    save(step_data, os);
-    os.close();
-}
-
 DIC_analysis_output DIC_analysis(const DIC_analysis_input &DIC_input) {
     typedef ROI2D::difference_type                              difference_type;
             
@@ -2925,7 +2642,9 @@ DIC_analysis_output DIC_analysis_sequential(const DIC_analysis_input &DIC_input,
         
         std::chrono::time_point<std::chrono::system_clock> start_rgdic = std::chrono::system_clock::now();
         
-        std::cout << "Debug:: " << current_seeds[0].x << " " << current_seeds[0].y << std::endl;
+        if (DIC_input.debug && !current_seeds.empty()) {
+            std::cout << "Debug:: " << current_seeds[0].x << " " << current_seeds[0].y << std::endl;
+        }
         auto disps = RGDIC_without_thread(DIC_input.imgs[ref_idx].get_gs(), 
                                DIC_input.imgs[cur_idx].get_gs(), 
                                roi_ref, 
@@ -3643,125 +3362,6 @@ Strain2D LS_strain(const Disp2D &disp, PERSPECTIVE perspective_type, double unit
     return { std::move(A_eyy), std::move(A_exy), std::move(A_exx), disp.get_roi().form_union(A_vp), disp.get_scalefactor() };
 }
 
-strain_analysis_input strain_analysis_input::load(std::ifstream &is) {
-    // Form empty strain_analysis_input then fill in values in accordance to how they are saved
-    strain_analysis_input strain_input;
-    
-    // Load DIC_input
-    strain_input.DIC_input = DIC_analysis_input::load(is);
-    
-    // Load DIC_output
-    strain_input.DIC_output = DIC_analysis_output::load(is);
-    
-    // Load subregion type
-    is.read(reinterpret_cast<char*>(&strain_input.subregion_type), std::streamsize(sizeof(SUBREGION)));
-    
-    // Load radius
-    is.read(reinterpret_cast<char*>(&strain_input.r), std::streamsize(sizeof(difference_type)));
-    
-    return strain_input;
-}
-
-strain_analysis_input strain_analysis_input::load(const std::string &filename) {
-    // Form stream
-    std::ifstream is(filename.c_str(), std::ios::in | std::ios::binary);
-    if (!is.is_open()) {
-        throw std::invalid_argument("Could not open " + filename + " for loading strain_analysis_input.");
-    }
-    
-    // Form strain_analysis_input using stream static factory method
-    auto strain_input = strain_analysis_input::load(is);
-    
-    // Close stream
-    is.close();
-    
-    return strain_input;
-}
-    
-void save(const strain_analysis_input &strain_input, std::ofstream &os) {    
-    typedef ROI2D::difference_type                              difference_type;
-    
-    // Save DIC_input -> DIC_output -> subregion_type -> r       
-   
-    save(strain_input.DIC_input, os);
-    
-    save(strain_input.DIC_output, os);    
-    
-    os.write(reinterpret_cast<const char*>(&strain_input.subregion_type), std::streamsize(sizeof(SUBREGION)));   
-    
-    os.write(reinterpret_cast<const char*>(&strain_input.r), std::streamsize(sizeof(difference_type)));   
-}
-
-void save(const strain_analysis_input &strain_input, const std::string &filename) {
-    // Form stream
-    std::ofstream os(filename.c_str(), std::ios::out | std::ios::binary | std::ios::trunc);
-    if (!os.is_open()) {
-        throw std::invalid_argument("Could not open " + filename + " for saving strain_analysis_input.");
-    }
-    
-    // Save strain_input into stream
-    save(strain_input, os);
-        
-    // Close stream
-    os.close();
-}
-
-strain_analysis_output strain_analysis_output::load(std::ifstream &is) {
-    // Form empty strain_analysis_output then fill in values in accordance to how they are saved
-    strain_analysis_output strain_output;
-    
-    // Load strains
-    difference_type num_strains = 0;
-    is.read(reinterpret_cast<char*>(&num_strains), std::streamsize(sizeof(difference_type)));
-    strain_output.strains.resize(num_strains);
-    for (auto &strain : strain_output.strains) {
-        strain = Strain2D::load(is);
-    }
-    
-    return strain_output;
-}
-
-strain_analysis_output strain_analysis_output::load(const std::string &filename) {
-    // Form stream
-    std::ifstream is(filename.c_str(), std::ios::in | std::ios::binary);
-    if (!is.is_open()) {
-        throw std::invalid_argument("Could not open " + filename + " for loading strain_analysis_output.");
-    }
-    
-    // Form strain_analysis_output using stream static factory method
-    auto strain_output = strain_analysis_output::load(is);
-    
-    // Close stream
-    is.close();
-    
-    return strain_output;
-}
-    
-void save(const strain_analysis_output &strain_output, std::ofstream &os) {    
-    typedef ROI2D::difference_type                              difference_type;
-    
-    // Save strains
-    difference_type num_strains = strain_output.strains.size();
-    os.write(reinterpret_cast<const char*>(&num_strains), std::streamsize(sizeof(difference_type)));    
-    for (const auto &strain : strain_output.strains) {
-        save(strain, os);
-    }    
-}
-
-void save(const strain_analysis_output &strain_output, const std::string &filename) {
-    // Form stream
-    std::ofstream os(filename.c_str(), std::ios::out | std::ios::binary | std::ios::trunc);
-    if (!os.is_open()) {
-        throw std::invalid_argument("Could not open " + filename + " for saving strain_analysis_output.");
-    }
-    
-    // Save strain_output into stream
-    save(strain_output, os);
-        
-    // Close stream
-    os.close();
-}
-
 strain_analysis_output strain_analysis(const strain_analysis_input &strain_input) {
     typedef ROI2D::difference_type                              difference_type;
 
@@ -3783,1472 +3383,4 @@ strain_analysis_output strain_analysis(const strain_analysis_input &strain_input
     return strain_output;
 }
 
-// Interface functions for viewing and saving data ---------------------------//
-namespace details {
-    cv::Mat cv_ncorr_data_over_img(const Image2D &img, 
-                                   const Data2D &data, 
-                                   double alpha, 
-                                   double min_data, 
-                                   double max_data, 
-                                   bool enable_colorbar = true, 
-                                   bool enable_axes = true,
-                                   bool enable_scalebar = true,
-                                   const std::string &units = "pixels",
-                                   double units_per_pixel = 1.0,
-                                   double num_units = -1.0,
-                                   double font_size = 1.0,
-                                   ROI2D::difference_type num_tick_marks = 11,
-                                   int colormap = cv::COLORMAP_JET) {
-        typedef ROI2D::difference_type                          difference_type;
-
-        // Data2D's used with Ncorr interface functions use the p1 direction for
-        // the y-axis (i.e.  pointing downwards) and the p2 direction for the 
-        // x-axis (i.e pointing rightwards) with the origin at the top-left. 
-        // This dictates how the axes are displayed and correspond to the data.
-        // For general Data2D's x and y axis have no assumptions about their 
-        // directions, so disable axes if general Data2D inputs are used.
-
-        if (alpha < 0 || alpha > 1) {
-            throw std::invalid_argument("alpha input for cv_ncorr_data_over_img() must be between 0 and 1.");
-        }
-
-        if (units_per_pixel <= 0) {
-            throw std::invalid_argument("units_per_pixel input for cv_ncorr_data_over_img() must be greater than 0.");
-        }
-
-        if (num_units != -1 && num_units <= 0) {
-            throw std::invalid_argument("num_units input for cv_ncorr_data_over_img() must be greater than 0.");
-        }
-        
-        if (font_size <= 0) {
-            throw std::invalid_argument("font_size input for cv_ncorr_data_over_img() must be greater than 0.");
-        }
-
-        if (num_tick_marks < 2) {
-            throw std::invalid_argument("num_tick_marks input for cv_ncorr_data_over_img() must be greater than or equal to 2.");
-        }    
-        
-        // Set up image first - must resize it to the data
-        auto A_img = img.get_gs();    
-        cv::Mat cv_img = get_cv_img(A_img, min(A_img), max(A_img));
-        cv::resize(cv_img, cv_img, { int(data.data_width()), int(data.data_height()) });
-        
-        // Set up data - must apply color map.
-        cv::Mat cv_data = get_cv_img(data.get_array(), min_data, max_data);
-        cv::applyColorMap(cv_data, cv_data, colormap);
-
-        // Set up plot by drawing the image with data over it
-        difference_type border = 20;        
-        difference_type plot_height = data.data_height() + 2*border;
-        difference_type plot_width = data.data_width() + 2*border;
-        cv::Mat cv_plot(plot_height, plot_width, CV_8UC3, cv::Vec3b(255,255,255)); // CV_8UC3 means 3-channel image        
-        for (difference_type p2 = border; p2 < data.data_width() + border; ++p2) {
-            difference_type p2_data = p2 - border;
-            for (difference_type p1 = border; p1 < data.data_height() + border; ++p1) {
-                difference_type p1_data = p1 - border;
-                
-                // Get data values
-                auto data_rgb = cv_data.at<cv::Vec3b>(p1_data,p2_data);
-                auto img_gs = cv_img.at<uchar>(p1_data,p2_data);
-
-                if (data.get_roi()(p1_data,p2_data)) {
-                    // Set data overlay if point is within ROI
-                    cv_plot.at<cv::Vec3b>(p1,p2) = cv::Vec3b(img_gs*(1-alpha) + alpha*data_rgb.val[0], 
-                                                             img_gs*(1-alpha) + alpha*data_rgb.val[1], 
-                                                             img_gs*(1-alpha) + alpha*data_rgb.val[2]);
-                } else {
-                    // Just draw background image
-                    cv_plot.at<cv::Vec3b>(p1,p2) = cv::Vec3b(img_gs, img_gs, img_gs);
-                }
-            }
-        } 
-        
-        if (enable_colorbar) {   
-            // Create color bar (if enabled) and horizontally concatenate it 
-            // with cv_plot. 
-                                    
-            // These are opencv parameters for text drawing
-            auto font_face = cv::FONT_HERSHEY_PLAIN;
-            int font_thickness = 1;
-            int baseline = 0;
-            
-            // Colorbar parameters - note that division by (num_tick_marks-1) is
-            // safe because num_tick_marks has been checked to ensure it is 
-            // greater than or equal to 2.
-            difference_type num_chars = 8; // number of chars printed for colorbar tick labels
-            difference_type text_offset_left = 5;
-            difference_type colorbar_width = 20;       
-            // Set colorbar background width based on widest string width
-            difference_type colorbar_bg_width = 0;
-            for (difference_type num_mark = 0; num_mark < num_tick_marks; ++num_mark) {
-                auto text_width = cv::getTextSize(std::to_string(double(num_tick_marks-num_mark-1)/(num_tick_marks-1) * (min_data-max_data) + max_data).substr(0,num_chars), font_face, 0.75*font_size, font_thickness, &baseline).width;
-                if (text_width > colorbar_bg_width) {
-                    colorbar_bg_width = text_width;
-                }
-            } 
-            colorbar_bg_width += colorbar_width + text_offset_left + border; // Finish width to account for spacing and offsets
-            
-            // Draw the color portion of colorbar first and then apply colormap
-            cv::Mat cv_colorbar(plot_height, colorbar_bg_width, cv::DataType<uchar>::type);            
-            for (difference_type p2 = 0; p2 < colorbar_width; ++p2) {
-                for (difference_type p1 = border; p1 < data.data_height() + border; ++p1) {
-                    // Applies value such that color will be max at the top of
-                    // the colorbar after the colormap is applied.
-                    cv_colorbar.at<uchar>(p1,p2) = double((data.data_height() + border - 1) - p1)/(data.data_height()-1) * 255;
-                }
-            } 
-            cv::applyColorMap(cv_colorbar, cv_colorbar, colormap);
-                        
-            // Paint region outside colorbar white
-            cv_colorbar(cv::Range(0,border),cv::Range::all()) = cv::Vec3b(255,255,255);
-            cv_colorbar(cv::Range(data.data_height()+border,data.data_height()+2*border),cv::Range::all()) = cv::Vec3b(255,255,255);
-            cv_colorbar(cv::Range::all(),cv::Range(colorbar_width,colorbar_bg_width)) = cv::Vec3b(255,255,255);
-            
-            // Paint black border inside colorbar
-            cv_colorbar(cv::Range(border, border+1),cv::Range(0,colorbar_width)) = cv::Vec3b(0,0,0);
-            cv_colorbar(cv::Range(data.data_height()+border-1,data.data_height()+border),cv::Range(0,colorbar_width)) = cv::Vec3b(0,0,0);
-            cv_colorbar(cv::Range(border,data.data_height()+border),cv::Range(0,1)) = cv::Vec3b(0,0,0);
-            cv_colorbar(cv::Range(border,data.data_height()+border),cv::Range(colorbar_width-1,colorbar_width)) = cv::Vec3b(0,0,0);
-                        
-            // Paint tick-marks and corresponding labels
-            difference_type tick_mark_width = 4;
-            for (difference_type num_mark = 0; num_mark < num_tick_marks; ++num_mark) {
-                difference_type p1 = (num_mark*(data.data_height()-1))/(num_tick_marks-1) + border;
-                
-                // Paint left mark, then right mark
-                cv_colorbar(cv::Range(p1, p1+1),cv::Range(0,tick_mark_width)) = cv::Vec3b(0,0,0);
-                cv_colorbar(cv::Range(p1, p1+1),cv::Range(colorbar_width-tick_mark_width,colorbar_width)) = cv::Vec3b(0,0,0);
-                
-                // Draw tick mark value
-                std::string tick_mark_label_str = std::to_string(num_mark*(min_data-max_data)/(num_tick_marks-1) + max_data).substr(0,num_chars);
-                auto text_size = cv::getTextSize(tick_mark_label_str, font_face, 0.75*font_size, font_thickness, &baseline);
-                cv::putText(cv_colorbar, 
-                            tick_mark_label_str, 
-                            cv::Point(colorbar_width + text_offset_left, num_mark*(data.data_height()-1)/(num_tick_marks-1)  + border + text_size.height/2),
-                            font_face, 
-                            0.75*font_size, 
-                            cv::Scalar::all(0), 
-                            font_thickness);
-            }           
-            
-            // Horizontally concatenate with cv_plot
-            cv::hconcat(cv_plot, cv_colorbar, cv_plot);
-        }
-        
-        if (enable_axes) {   
-            // axes_length is the longest dimension of the background axes
-            difference_type axes_length = 0.25 * std::min(data.data_height(), data.data_width());
-            
-            // background axes is painted black
-            cv::Point bg_axes_pts[1][8];
-            bg_axes_pts[0][0] = cv::Point(border,                   border);
-            bg_axes_pts[0][1] = cv::Point(border,                   border+axes_length);
-            bg_axes_pts[0][2] = cv::Point(border+0.20*axes_length,  border+0.80*axes_length);
-            bg_axes_pts[0][3] = cv::Point(border+0.10*axes_length,  border+0.80*axes_length);
-            bg_axes_pts[0][4] = cv::Point(border+0.10*axes_length,  border+0.10*axes_length);
-            bg_axes_pts[0][5] = cv::Point(border+0.80*axes_length,  border+0.10*axes_length);
-            bg_axes_pts[0][6] = cv::Point(border+0.80*axes_length,  border+0.20*axes_length);
-            bg_axes_pts[0][7] = cv::Point(border+axes_length,       border);
-            const cv::Point* pts1[1] = { bg_axes_pts[0] };
-            int npts1[] = { 8 };
-            cv::fillPoly(cv_plot, pts1, npts1, 1, cv::Scalar(255,255,255));
-            
-            // foreground axes is painted white over the black background axes
-            cv::Point fg_axes_pts[1][8];
-            fg_axes_pts[0][0] = cv::Point(border,                   border);
-            fg_axes_pts[0][1] = cv::Point(border,                   border+0.95*axes_length);
-            fg_axes_pts[0][2] = cv::Point(border+0.125*axes_length, border+0.825*axes_length);
-            fg_axes_pts[0][3] = cv::Point(border+0.070*axes_length, border+0.825*axes_length);
-            fg_axes_pts[0][4] = cv::Point(border+0.070*axes_length, border+0.070*axes_length);
-            fg_axes_pts[0][5] = cv::Point(border+0.825*axes_length, border+0.070*axes_length);
-            fg_axes_pts[0][6] = cv::Point(border+0.825*axes_length, border+0.125*axes_length);
-            fg_axes_pts[0][7] = cv::Point(border+0.95*axes_length,  border);
-            const cv::Point* pts2[1] = { fg_axes_pts[0] };
-            int npts2[] = { 8 };
-            cv::fillPoly(cv_plot, pts2, npts2, 1, cv::Scalar(0,0,0));
-            
-            // Draw Y and X labels
-            difference_type label_border = 5;
-            difference_type label_offset = 5;
-            double label_alpha = 0.5;
-            
-            // These are opencv parameters for text drawing            
-            auto font_face = cv::FONT_HERSHEY_PLAIN;
-            int font_thickness = 1;
-            int baseline = 0;                  
-            
-            // Draw "Y" label
-            std::string y_str = "Y";
-            auto text_size_y = cv::getTextSize(y_str, font_face, font_size, font_thickness, &baseline);
-            for (difference_type p2 = border+label_offset; p2 < border+label_offset+2.0*label_border+text_size_y.width; ++p2) {
-                for (difference_type p1 = border+axes_length+label_offset; p1 < border+axes_length+label_offset+2.0*label_border+text_size_y.height; ++p1) {
-                    if (p1 >= 0 && p1 < cv_plot.rows && 
-                        p2 >= 0 && p2 < cv_plot.cols) {
-                        auto cv_plot_rgb = cv_plot.at<cv::Vec3b>(p1,p2);
-                
-                        cv_plot.at<cv::Vec3b>(p1,p2) = cv::Vec3b(label_alpha*cv_plot_rgb.val[0], 
-                                                                 label_alpha*cv_plot_rgb.val[1], 
-                                                                 label_alpha*cv_plot_rgb.val[2]);
-                    }
-                }
-            }
-            cv::putText(cv_plot, 
-                        y_str, 
-                        cv::Point(border+label_offset+label_border, border+axes_length+label_offset+label_border+text_size_y.height),
-                        font_face, 
-                        font_size, 
-                        cv::Scalar::all(255), 
-                        font_thickness);
-            
-            // Draw "X" label
-            std::string x_str = "X";
-            auto text_size_x = cv::getTextSize(x_str, font_face, font_size, font_thickness, &baseline);
-            for (difference_type p2 = border+axes_length+label_offset; p2 < border+axes_length+label_offset+2.0*label_border+text_size_x.width; ++p2) {
-                for (difference_type p1 = border+label_offset; p1 < border+label_offset+2.0*label_border+text_size_x.height; ++p1) {
-                    if (p1 >= 0 && p1 < cv_plot.rows && 
-                        p2 >= 0 && p2 < cv_plot.cols) {
-                        auto cv_plot_rgb = cv_plot.at<cv::Vec3b>(p1,p2);
-                
-                        cv_plot.at<cv::Vec3b>(p1,p2) = cv::Vec3b(label_alpha*cv_plot_rgb.val[0], 
-                                                                 label_alpha*cv_plot_rgb.val[1], 
-                                                                 label_alpha*cv_plot_rgb.val[2]);
-                    }
-                }
-            }
-            cv::putText(cv_plot, 
-                        x_str, 
-                        cv::Point(border+axes_length+label_offset+label_border, border+label_offset+label_border+text_size_x.height),
-                        font_face, 
-                        font_size, 
-                        cv::Scalar::all(255), 
-                        font_thickness);      
-        }
-                
-        if (enable_scalebar) {   
-            // Set scalebar dimensions
-            difference_type scalebar_width = num_units == -1 ? data.data_width()/2 : num_units/units_per_pixel/data.get_scalefactor();
-            difference_type scalebar_height = 5;
-            
-            // Set num_units if it isn't -1
-            if (num_units == -1) {
-                num_units = scalebar_width*units_per_pixel*data.get_scalefactor();
-            }
-            
-            // Form text first so we can use its size in determining the scalebar
-            // bg height.
-            difference_type decimal_places = 2;
-            std::stringstream ss_scalebar;
-            ss_scalebar << std::fixed << std::setprecision(std::numeric_limits<double>::digits10) << num_units;
-            std::string scalebar_str_unformatted = ss_scalebar.str();
-            auto idx_period = scalebar_str_unformatted.find(".");
-            std::string scalebar_str = idx_period == std::string::npos ? scalebar_str_unformatted : scalebar_str_unformatted.substr(0, idx_period + decimal_places + 1) + " " + units;
-                        
-            // These are opencv parameters for text drawing            
-            auto font_face = cv::FONT_HERSHEY_PLAIN;
-            int font_thickness = 1;
-            int baseline = 0;            
-            auto text_size = cv::getTextSize(scalebar_str, font_face, font_size, font_thickness, &baseline);   
-            
-            // Draw background first
-            difference_type scalebar_bg_border = 10;
-            difference_type scalebar_bg_width = scalebar_width + 2*scalebar_bg_border;
-            difference_type scalebar_bg_height = scalebar_height + 3*scalebar_bg_border + text_size.height;
-            difference_type scalebar_bg_offset = 10;
-            double scalebar_bg_alpha = 0.5;
-            for (difference_type p2 = border+scalebar_bg_offset; p2 < border+scalebar_bg_offset+scalebar_bg_width; ++p2) {
-                for (difference_type p1 = border+data.data_height()-scalebar_bg_offset-scalebar_bg_height; p1 < border+data.data_height()-scalebar_bg_offset; ++p1) {                   
-                    if (p1 >= 0 && p1 < cv_plot.rows && 
-                        p2 >= 0 && p2 < cv_plot.cols) {
-                        auto cv_plot_rgb = cv_plot.at<cv::Vec3b>(p1,p2);
-
-                        cv_plot.at<cv::Vec3b>(p1,p2) = cv::Vec3b(scalebar_bg_alpha*cv_plot_rgb.val[0], 
-                                                                 scalebar_bg_alpha*cv_plot_rgb.val[1], 
-                                                                 scalebar_bg_alpha*cv_plot_rgb.val[2]);
-                    }
-                }
-            }
-                        
-            // Draw scalebar
-            for (difference_type p2 = border+scalebar_bg_offset+scalebar_bg_border; p2 < border+scalebar_bg_offset+scalebar_bg_border+scalebar_width; ++p2) {
-                for (difference_type p1 = border+data.data_height()-scalebar_bg_offset-scalebar_bg_border-scalebar_height; p1 < border+data.data_height()-scalebar_bg_offset-scalebar_bg_border; ++p1) {
-                    if (p1 >= 0 && p1 < cv_plot.rows && 
-                        p2 >= 0 && p2 < cv_plot.cols) {
-                        cv_plot.at<cv::Vec3b>(p1,p2) = cv::Vec3b(255,255,255);
-                    }
-                }
-            }
-            
-            // Draw scalebar text          
-            cv::putText(cv_plot, 
-                        scalebar_str, 
-                        cv::Point(border+scalebar_bg_offset+scalebar_bg_width/2.0-text_size.width/2.0, border+data.data_height()-scalebar_bg_offset-2*scalebar_bg_border-scalebar_height),
-                        font_face, 
-                        font_size, 
-                        cv::Scalar::all(255), 
-                        font_thickness);    
-        }
-        
-        return cv_plot;
-    }
-}
-
-void imshow_ncorr_data_over_img(const Image2D &img, const Data2D &data, ROI2D::difference_type delay) {          
-    // Set parameters 
-    double min_data = 0, max_data = 0; 
-    Array2D<double> data_values = data.get_array()(data.get_roi().get_mask());
-    if (!data_values.empty()) {
-        min_data = prctile(data_values,0.01);
-        max_data = prctile(data_values,0.99);
-    }    
-    
-    // Get cv style matrix and show it - disable scalebar for simplicity since
-    // that requires units and unit conversion parameters and this function is
-    // mainly used for debugging purposes.
-    auto cv_img = details::cv_ncorr_data_over_img(img, 
-                                                  data, 
-                                                  0.5,          // Alpha
-                                                  min_data, 
-                                                  max_data, 
-                                                  true,         // enable_colorbar
-                                                  true,         // enable_axes
-                                                  false);       // enable_scalebar
-    cv::imshow("Ncorr data", cv_img);
-    delay == -1 ? cv::waitKey() : cv::waitKey(delay);
-}
-
-void save_ncorr_data_over_img(const std::string &filename, 
-                              const Image2D &img,
-                              const Data2D &data,
-                              double alpha,
-                              double min_data,
-                              double max_data,
-                              bool enable_colorbar,
-                              bool enable_axes,
-                              bool enable_scalebar,
-                              const std::string &units,
-                              double units_per_pixel,
-                              double num_units,
-                              double font_size,
-                              ROI2D::difference_type num_tick_marks,
-                              int colormap) {   
-    // Get cv style matrix and save it
-    auto cv_data_img = details::cv_ncorr_data_over_img(img, 
-                                                       data, 
-                                                       alpha,
-                                                       min_data, 
-                                                       max_data, 
-                                                       enable_colorbar,
-                                                       enable_axes,
-                                                       enable_scalebar,
-                                                       units,
-                                                       units_per_pixel,
-                                                       num_units,
-                                                       font_size,
-                                                       num_tick_marks,
-                                                       colormap);
-    cv::imwrite(filename, cv_data_img);
-}
-
-void save_ncorr_data_over_img_video(const std::string &filename, 
-                                    const std::vector<Image2D> &imgs,
-                                    const std::vector<Data2D> &data,
-                                    double alpha,
-                                    double fps,
-                                    double min_data,
-                                    double max_data,
-                                    bool enable_colorbar,
-                                    bool enable_axes,
-                                    bool enable_scalebar,
-                                    const std::string &units,
-                                    double units_per_pixel,
-                                    double num_units,
-                                    double font_size,
-                                    ROI2D::difference_type num_tick_marks,
-                                    int colormap,
-                                    double end_delay,
-                                    int fourcc) {
-    typedef ROI2D::difference_type                          difference_type;
-    
-    // Note that if only 1 image was input, interpret this as using the same 
-    // image for each Data2D. If not, use the img the data corresponds to.
-    
-    if (imgs.size() != 1 && imgs.size() != data.size()) {        
-        // Number of images must either be 1, or equal to the number of input data
-        throw std::invalid_argument("Number of images used in save_data_over_img_video() must either be 1, or equal to the number of input data.");
-    }
-    
-    if (data.empty()) {
-        throw std::invalid_argument("Number of Data2D used in save_data_over_img_video() must be greater than or equal to 1.");
-    }
-    
-    // All data must have the same size
-    for (difference_type data_idx = 1; data_idx < difference_type(data.size()); ++data_idx) {
-        if (data[data_idx].data_height() != data.front().data_height() ||
-            data[data_idx].data_width() != data.front().data_width()) {
-            throw std::invalid_argument("Attempted to use save_data_over_img_video() with data of differing sizes. All data must be the same size.");
-        }
-    }
-    
-    if (fps <= 0) {
-        throw std::invalid_argument("fps input for save_data_over_img_video() must be greater than 0.");
-    }
-    
-    if (end_delay < 0) {
-        throw std::invalid_argument("end_delay input for save_data_over_img_video() must be greater than or equal to 0.");
-    }    
-    
-    // All other parameters will be checked when calling cv_ncorr_data_over_img()
-
-    std::cout << std::endl << "Saving video: " << filename << "..." << std::endl;
-
-    // Initialize video    
-    cv::VideoWriter output_video;  
-    
-    // Cycle over data and save
-    for (difference_type data_idx = 0; data_idx < difference_type(data.size()); ++data_idx) {
-        std::cout << "Frame " << data_idx+1 << " of " << data.size() << "." << std::endl;
-        
-        auto cv_data_img = details::cv_ncorr_data_over_img(imgs.size() == 1 ? imgs.front() : imgs[data_idx], 
-                                                           data[data_idx], 
-                                                           alpha,
-                                                           min_data, 
-                                                           max_data, 
-                                                           enable_colorbar,
-                                                           enable_axes,
-                                                           enable_scalebar,
-                                                           units,
-                                                           units_per_pixel,
-                                                           num_units,
-                                                           font_size,
-                                                           num_tick_marks,
-                                                           colormap);
-        
-        if (data_idx == 0) {
-            // Open video file - do this here because the size of the cv_data_img
-            // depends on its inputs and the size is needed to open the video.
-            output_video.open(filename, 
-                              fourcc, 
-                              fps, 
-                              { cv_data_img.cols, cv_data_img.rows }, 
-                              true);            
-
-            if (!output_video.isOpened()) {
-                throw std::invalid_argument("Cannot open video file: " + filename + " for save_data_over_img_video().");
-            }
-        }
-        
-        output_video << cv_data_img;
-    }
-
-    // Add the last frame again multiple times to provide a delay
-    for (difference_type idx = 0; idx < fps*end_delay; ++idx) {
-        output_video << details::cv_ncorr_data_over_img(imgs.size() == 1 ? imgs.front() : imgs.back(), 
-                                                        data.back(), 
-                                                        alpha,
-                                                        min_data, 
-                                                        max_data, 
-                                                        enable_colorbar,
-                                                        enable_axes,
-                                                        enable_scalebar,
-                                                        units,
-                                                        units_per_pixel,
-                                                        num_units,
-                                                        font_size,
-                                                        num_tick_marks,
-                                                        colormap);
-    }
-}
-
-void save_DIC_video(const std::string &filename, 
-                    const DIC_analysis_input &DIC_input, 
-                    const DIC_analysis_output &DIC_output, 
-                    DISP disp_type,
-                    double alpha,
-                    double fps,
-                    double min_disp,
-                    double max_disp,
-                    bool enable_colorbar,
-                    bool enable_axes,
-                    bool enable_scalebar,
-                    double num_units,
-                    double font_size,
-                    ROI2D::difference_type num_tick_marks,
-                    int colormap,
-                    double end_delay,
-                    int fourcc) {
-    typedef ROI2D::difference_type                              difference_type;
-                        
-    // Use get_disp() function to obtain the specified displacement field
-    std::function<const Data2D&(const Disp2D&)> get_disp;
-    switch (disp_type) {
-        case DISP::V : 
-            get_disp = &Disp2D::get_v;
-            break;
-        case DISP::U : 
-            get_disp = &Disp2D::get_u;
-            break;            
-    }   
-                
-    // Set min and max if they are NaN. They are set so that all data fits 
-    // approximately between these bounds. This assumes maxima occur at the
-    // first or last displacement plot for simplicity.
-    Array2D<double> data_values_first = get_disp(DIC_output.disps.front()).get_array()(DIC_output.disps.front().get_roi().get_mask());
-    Array2D<double> data_values_last = get_disp(DIC_output.disps.back()).get_array()(DIC_output.disps.back().get_roi().get_mask());
-    if (std::isnan(min_disp) && !data_values_first.empty() && !data_values_last.empty()) {
-        min_disp = std::min(prctile(data_values_first,0.01), prctile(data_values_last,0.01));
-    }
-    if (std::isnan(max_disp) && !data_values_first.empty() && !data_values_last.empty()) {
-        max_disp = std::max(prctile(data_values_first,0.99), prctile(data_values_last,0.99));
-    }
-    
-    // Get data plots
-    std::vector<Data2D> data;
-    for (difference_type disp_idx = 0; disp_idx < difference_type(DIC_output.disps.size()); ++disp_idx) {
-        data.push_back(get_disp(DIC_output.disps[disp_idx]));
-    }
-    
-    // Get imgs - this depends on perspective
-    std::vector<Image2D> imgs;
-    switch (DIC_output.perspective_type) {
-        case PERSPECTIVE::LAGRANGIAN : 
-            imgs.push_back(DIC_input.imgs.front());
-            break;
-        case PERSPECTIVE::EULERIAN : 
-            for (difference_type img_idx = 1; img_idx < difference_type(DIC_input.imgs.size()); ++img_idx) { // Starts from 1
-                imgs.push_back(DIC_input.imgs[img_idx]);
-            }
-            break;            
-    }   
-    
-    // Save the video
-    save_ncorr_data_over_img_video(filename, 
-                                   imgs, 
-                                   data, 
-                                   alpha, 
-                                   fps, 
-                                   min_disp, 
-                                   max_disp,  
-                                   enable_colorbar, 
-                                   enable_axes, 
-                                   enable_scalebar, 
-                                   DIC_output.units, 
-                                   DIC_output.units_per_pixel,
-                                   num_units,
-                                   font_size,
-                                   num_tick_marks,
-                                   colormap,
-                                   end_delay,
-                                   fourcc);
-}
-
-void save_strain_video(const std::string &filename, 
-                       const strain_analysis_input &strain_input, 
-                       const strain_analysis_output &strain_output, 
-                       STRAIN strain_type,
-                       double alpha,
-                       double fps, 
-                       double min_strain,
-                       double max_strain,
-                       bool enable_colorbar,
-                       bool enable_axes,
-                       bool enable_scalebar,
-                       double num_units,
-                       double font_size,
-                       ROI2D::difference_type num_tick_marks,
-                       int colormap,
-                       double end_delay,
-                       int fourcc) {
-    typedef ROI2D::difference_type                              difference_type;
-                        
-    // Use get_strain function to obtain the specified strain field
-    std::function<const Data2D&(const Strain2D&)> get_strain;
-    switch (strain_type) {
-        case STRAIN::EYY : 
-            get_strain = &Strain2D::get_eyy;
-            break;
-        case STRAIN::EXY : 
-            get_strain = &Strain2D::get_exy;
-            break;
-        case STRAIN::EXX : 
-            get_strain = &Strain2D::get_exx;
-            break;       
-    }   
-        
-    // Set min and max if they are NaN. They are set so that all data fits 
-    // approximately between these bounds This assumes maxima occur at the
-    // first or last strain plot for simplicity.
-    Array2D<double> data_values_first = get_strain(strain_output.strains.front()).get_array()(strain_output.strains.front().get_roi().get_mask());
-    Array2D<double> data_values_last = get_strain(strain_output.strains.back()).get_array()(strain_output.strains.back().get_roi().get_mask());
-    if (std::isnan(min_strain) && !data_values_first.empty() && !data_values_last.empty()) {
-        min_strain = std::min(prctile(data_values_first,0.01), prctile(data_values_last,0.01));
-    }
-    if (std::isnan(max_strain) && !data_values_first.empty() && !data_values_last.empty()) {
-        max_strain = std::max(prctile(data_values_first,0.99), prctile(data_values_last,0.99));
-    }
-            
-    // Get data plots
-    std::vector<Data2D> data;
-    for (difference_type strain_idx = 0; strain_idx < difference_type(strain_output.strains.size()); ++strain_idx) {
-        data.push_back(get_strain(strain_output.strains[strain_idx]));
-    }
-    
-    // Get imgs - this depends on perspective
-    std::vector<Image2D> imgs;
-    switch (strain_input.DIC_output.perspective_type) {
-        case PERSPECTIVE::LAGRANGIAN : 
-            imgs.push_back(strain_input.DIC_input.imgs.front());
-            break;
-        case PERSPECTIVE::EULERIAN : 
-            for (difference_type img_idx = 1; img_idx < difference_type(strain_input.DIC_input.imgs.size()); ++img_idx) { // Starts from 1
-                imgs.push_back(strain_input.DIC_input.imgs[img_idx]);
-            }
-            break;            
-    }   
-        
-    // Save the video
-    save_ncorr_data_over_img_video(filename, 
-                                   imgs, 
-                                   data, 
-                                   alpha, 
-                                   fps, 
-                                   min_strain, 
-                                   max_strain, 
-                                   enable_colorbar, 
-                                   enable_axes, 
-                                   enable_scalebar, 
-                                   strain_input.DIC_output.units, 
-                                   strain_input.DIC_output.units_per_pixel,
-                                   num_units,
-                                   font_size,
-                                   num_tick_marks,
-                                   colormap,
-                                   end_delay, 
-                                   fourcc);
-}
-
-
-// Seed-based parallel DIC analysis implementation ---------------------------//
-
-// Compute displacements using precomputed seed parameters
-Disp2D compute_displacements(
-    const details::subregion_nloptimizer &sr_nloptimizer,
-    const ROI2D& roi_reduced,
-    const SeedParams& seedparams,
-    ROI2D::difference_type scalefactor,
-    double cutoff_corrcoef,
-    ROI2D::difference_type region_idx,
-    bool debug
-) {
-
-    // Perform RGDIC with manual seed (reliability-guided expansion from seed point)
-    auto H = roi_reduced.height();
-    auto W = roi_reduced.width();
-    
-    // cutoff_delta_disp is a cutoff for the allowable change in displacement 
-    // between analyzed points. Generally if a point is analyzed incorrectly,
-    // it will result in a large change in displacement, so just filter these
-    // points out.
-    double cutoff_delta_disp = scalefactor;
-
-    // -----------------------------------------------------------------------//
-    // Perform reliability guided DIC without threading ----------------------//
-    // -----------------------------------------------------------------------//    
-    // Initialize buffers - set A_ap to all points in ROI
-    // Create arrays for displacement computation
-    Array2D<double> A_v(H, W);
-    Array2D<double> A_u(H, W);
-    Array2D<double> A_cc(H, W);
-    Array2D<bool> A_ap(roi_reduced.get_mask());                // Active Points
-    Array2D<bool> A_vp(H, W); // Valid Points
-    
-    // Cycle over regions and perform RGDIC sequentially
-    Array2D<double> params_buf(10, 1); // buffer for nloptimizer
-    
-    // Get seed params for queue
-    // Must match SeedParams::to_array() convention:
-    // params = {p1(y), p2(x), v, u, dv_dp1(dv_dy), dv_dp2(dv_dx), du_dp1(du_dy), du_dp2(du_dx), corr_coef, diff_norm}
-    Array2D<double> seed_params(10, 1);
-    seed_params(0) = seedparams.y;      // p1 (row)
-    seed_params(1) = seedparams.x;      // p2 (col)
-    seed_params(2) = seedparams.v;
-    seed_params(3) = seedparams.u;
-    seed_params(4) = seedparams.dv_dy;  // dv_dp1
-    seed_params(5) = seedparams.dv_dx;  // dv_dp2
-    seed_params(6) = seedparams.du_dy;  // du_dp1
-    seed_params(7) = seedparams.du_dx;  // du_dp2
-    seed_params(8) = seedparams.corrcoef;
-    seed_params(9) = 0;
-    
-    if (!seed_params.empty()) {    
-        A_ap(seed_params(0) / scalefactor, seed_params(1) / scalefactor) = false;
-        
-        // Form queue - make it a priority queue such that the lowest 
-        // correlation coefficient values are processed first.
-        auto comp = [](const Array2D<double> &a, const Array2D<double> &b ) { return a(8) > b(8); };
-        std::priority_queue<Array2D<double>, std::vector<Array2D<double>>, std::function<bool(const Array2D<double>&,const Array2D<double>&)>> queue(comp);
-
-        // Perform flood fill around seed  
-        queue.push(seed_params);                            
-        while (!queue.empty()) {	                    
-            // 1) pop info from queue                
-            auto queue_params = std::move(queue.top()); queue.pop();
-
-            // 2) Validate and store information
-            A_vp(queue_params(0) / scalefactor,  queue_params(1) / scalefactor) = true;
-            A_v(queue_params(0) / scalefactor,  queue_params(1) / scalefactor) = queue_params(2);
-            A_u(queue_params(0) / scalefactor,  queue_params(1) / scalefactor) = queue_params(3);
-            A_cc(queue_params(0) / scalefactor,  queue_params(1) / scalefactor) = queue_params(8);
-                                    
-            // 3) Analyze four surrounding points - up, down, left right;
-            details::analyze_point(queue_params, -scalefactor, 0, roi_reduced.get_nlinfo(region_idx), scalefactor, sr_nloptimizer, cutoff_corrcoef, cutoff_delta_disp, queue, A_ap, params_buf);
-            details::analyze_point(queue_params,  scalefactor, 0, roi_reduced.get_nlinfo(region_idx), scalefactor, sr_nloptimizer, cutoff_corrcoef, cutoff_delta_disp, queue, A_ap, params_buf);
-            details::analyze_point(queue_params, 0, -scalefactor, roi_reduced.get_nlinfo(region_idx), scalefactor, sr_nloptimizer, cutoff_corrcoef, cutoff_delta_disp, queue, A_ap, params_buf);
-            details::analyze_point(queue_params, 0,  scalefactor, roi_reduced.get_nlinfo(region_idx), scalefactor, sr_nloptimizer, cutoff_corrcoef, cutoff_delta_disp, queue, A_ap, params_buf);
-        }    
-    }
-    
-    // Go back over plot and clear any datapoints that have large displacement
-    // jumps. Not all are cleared through the RGDIC algorithm because of the 
-    // way the search is conducted.
-    auto A_vp_buf = A_vp; // Make copy of valid points since it gets overwritten
-    for (ROI2D::difference_type region_idx = 0; region_idx < roi_reduced.size_regions(); ++region_idx) {
-        for (ROI2D::difference_type nl_idx = 0; nl_idx < roi_reduced.get_nlinfo(region_idx).nodelist.width(); ++nl_idx) {
-            ROI2D::difference_type p2 = nl_idx + roi_reduced.get_nlinfo(region_idx).left_nl;
-            for (ROI2D::difference_type np_idx = 0; np_idx < roi_reduced.get_nlinfo(region_idx).noderange(nl_idx); np_idx += 2) {
-                ROI2D::difference_type np_top = roi_reduced.get_nlinfo(region_idx).nodelist(np_idx,nl_idx);
-                ROI2D::difference_type np_bottom = roi_reduced.get_nlinfo(region_idx).nodelist(np_idx + 1,nl_idx);
-                for (ROI2D::difference_type p1 = np_top; p1 <= np_bottom; ++p1) {
-                    if ((roi_reduced.get_nlinfo(region_idx).in_nlinfo(p1-1,p2) && A_vp_buf(p1-1,p2) && std::sqrt(std::pow(A_v(p1-1,p2) - A_v(p1,p2),2) + std::pow(A_u(p1-1,p2) - A_u(p1,p2),2)) > cutoff_delta_disp) ||
-                        (roi_reduced.get_nlinfo(region_idx).in_nlinfo(p1+1,p2) && A_vp_buf(p1+1,p2) && std::sqrt(std::pow(A_v(p1+1,p2) - A_v(p1,p2),2) + std::pow(A_u(p1+1,p2) - A_u(p1,p2),2)) > cutoff_delta_disp) ||
-                        (roi_reduced.get_nlinfo(region_idx).in_nlinfo(p1,p2-1) && A_vp_buf(p1,p2-1) && std::sqrt(std::pow(A_v(p1,p2-1) - A_v(p1,p2),2) + std::pow(A_u(p1,p2-1) - A_u(p1,p2),2)) > cutoff_delta_disp) ||
-                        (roi_reduced.get_nlinfo(region_idx).in_nlinfo(p1,p2+1) && A_vp_buf(p1,p2+1) && std::sqrt(std::pow(A_v(p1,p2+1) - A_v(p1,p2),2) + std::pow(A_u(p1,p2+1) - A_u(p1,p2),2)) > cutoff_delta_disp)) {
-                        A_vp(p1,p2) = false;
-                    }
-                }
-            }
-        }
-    }            
-    
-    // Must form union with valid points
-    auto roi_valid = roi_reduced.form_union(A_vp); 
-    if (debug) {
-        std::cout << "RGDIC completed with manual seed" << std::endl;
-    }   
-    return Disp2D(std::move(A_v), std::move(A_u), std::move(A_cc), roi_valid, scalefactor);
-}
-
-// Compute seed parameters for all frames with ROI updates
-std::vector<SeedComputationData> compute_only_seed_points(
-    const Array2D<double>& A_ref,
-    const std::vector<Array2D<double>>& A_curs,
-    const ROI2D& roi,
-    ROI2D::difference_type scalefactor,
-    INTERP interp_type,
-    SUBREGION subregion_type,
-    ROI2D::difference_type r,
-    const std::vector<SeedParams>& seeds_by_region,
-    double cutoff_corrcoef,
-    ROI2D::difference_type region_idx,
-    bool debug
-) {
-
-    std::vector<SeedComputationData> selected_data;
-
-    if (debug) {
-        std::cout << "\n=== Compute seed parameters for all frames with ROI updates ===> ";
-        std::cout << "\n starting with " << seeds_by_region.size() << " seeds" << std::endl;
-    }
-    
-    // Start with initial reference and ROI
-    Array2D<double> A_ref_current = A_ref;
-    ROI2D roi_current = roi;
-    std::vector<SeedParams> seeds_current = seeds_by_region;
-    
-    size_t idx = 0;
-    while (idx < A_curs.size()) {
-        const Array2D<double>& A_cur = A_curs[idx];
-        
-        if (debug) {
-            std::cout << "\n=== Seeding analyzing frame " << (idx + 1) << " ===> ";
-        }
-
-        // Get subregion nonlinear optimizer
-        auto sr_nloptimizer = details::subregion_nloptimizer(A_ref_current, A_cur, roi_current, scalefactor, interp_type, subregion_type, r);
-        
-        // Analyze seeds for this frame
-        SeedAnalysisResult seed_results = analyze_seeds(
-            sr_nloptimizer,
-            A_ref_current,
-            roi_current,
-            seeds_current,
-            r,
-            50,     // cutoff_iteration
-            0.1,    // cutoff_max_diffnorm
-            0.5     // cutoff_max_corrcoef
-        );
-        
-        if (seed_results.success) {
-            // Seed analysis successful - store data
-            if (debug) {
-                std::cout << "Successful!" << std::endl;
-            }
-            
-            selected_data.emplace_back(roi_current, seed_results.seeds, sr_nloptimizer);
-            idx++;
-        } else {
-            // Seed analysis failed - update reference and propagate seeds
-            if (debug) {
-                std::cout << "Failed! Updating reference ===>" << std::endl;
-            }
-
-            auto A_prev = A_curs[idx-1];
-
-            // Get subregion nonlinear optimizer from the last added selected_data
-            auto prev_sr_nloptimizer = selected_data.back().sr_nloptimizer;
-            auto prev_seedparams = selected_data.back().seed_params_by_region;
-            auto prev_roi = selected_data.back().roi;
-            
-            // Compute displacements between at the previous frame
-            auto disps = compute_displacements(
-                prev_sr_nloptimizer,
-                prev_roi.reduce(scalefactor),
-                prev_seedparams[region_idx],
-                scalefactor,
-                cutoff_corrcoef,
-                region_idx,
-                debug
-            );
-            
-            // Update reference image and ROI
-            A_ref_current = A_prev;
-            roi_current = update(prev_roi, disps, interp_type);
-            
-            // Propagate seeds to track material deformation
-            seeds_current = propagate_seeds(prev_seedparams, scalefactor);
-            
-            if (debug) {
-                std::cout << "Reference updated, seeds propagated <====" << std::endl;
-            }
-        }
-    }
-    
-    if (debug) {
-        std::cout << "\nComputed seed parameters for " << selected_data.size() << " frames" << std::endl;
-    }
-    
-    return selected_data;
-}
-
-// Update seed positions based on displacement (seed propagation)
-std::vector<SeedParams> propagate_seeds(const std::vector<SeedParams>& seeds, ROI2D::difference_type spacing) {
-    std::vector<SeedParams> propagated_seeds;
-    propagated_seeds.reserve(seeds.size());
-    
-    for (const auto& seed : seeds) {
-        SeedParams new_seed = seed;
-        
-        // Update position based on displacement, rounding to spaced grid
-        // This matches the Matlab logic: pos + round(u/(spacing+1))*(spacing+1)
-        new_seed.x = seed.x + std::round(seed.u / (spacing + 1)) * (spacing + 1);
-        new_seed.y = seed.y + std::round(seed.v / (spacing + 1)) * (spacing + 1);
-        
-        propagated_seeds.push_back(new_seed);
-    }
-    
-    return propagated_seeds;
-}
-
-// Lightweight seed analysis for predicting reference updates
-SeedAnalysisResult analyze_seeds(
-    const details::subregion_nloptimizer &sr_nloptimizer,
-    const Array2D<double>& ref_gs,
-    const ROI2D& roi,
-    const std::vector<SeedParams>& seed_positions,
-    ROI2D::difference_type radius,
-    int cutoff_iteration,
-    double cutoff_max_diffnorm,
-    double cutoff_max_corrcoef,
-    bool debug
-) {
-    
-    SeedAnalysisResult result;
-    result.seeds.reserve(seed_positions.size());
-    result.convergence.reserve(seed_positions.size());
-    result.success = true;
-    
-    // Analyze each seed point using existing subregion optimizer
-    int region_idx = 0;
-    for (const auto& seed_pos : seed_positions) {
-        // Check if seed is within valid bounds
-        if (seed_pos.x < radius || seed_pos.x >= ref_gs.width() - radius ||
-            seed_pos.y < radius || seed_pos.y >= ref_gs.height() - radius) {
-            result.success = false;
-            continue;
-        }
-        
-        // Check if seed is in ROI
-        if (!roi.get_mask()(seed_pos.y, seed_pos.x)) {
-            result.success = false;
-            continue;
-        }
-        
-        // Initialize parameters for optimizer
-        // params = {p1, p2, v, u, dv_dp1, dv_dp2, du_dp1, du_dp2, corr_coef, diff_norm}
-        Array2D<double> params_init(10, 1);
-        params_init(0) = seed_pos.y;  // p1 in full resolution
-        params_init(1) = seed_pos.x;  // p2 in full resolution
-        params_init(2) = 0.0;  // v
-        params_init(3) = 0.0;  // u
-        params_init(4) = 0.0;  // dv_dp1
-        params_init(5) = 0.0;  // dv_dp2
-        params_init(6) = 0.0;  // du_dp1
-        params_init(7) = 0.0;  // du_dp2
-        params_init(8) = 0.0;  // corr_coef
-        params_init(9) = 0.0;  // diff_norm
-        
-        // Run IC-GN optimization via global() method
-        auto result_pair = sr_nloptimizer.global(params_init);
-        const auto& params_guess = result_pair.first;
-        
-        if (!result_pair.second) {
-            if (debug) {
-                std::cout << region_idx << ": Seed analysis failed at seed " << seed_pos.x << ", " << seed_pos.y << "=> global failed" << std::endl;
-            }
-            result.success = false;
-            continue;
-        }
-
-        // Run iterative-search
-        auto result_iter = sr_nloptimizer(params_guess);
-        const auto& params_result = result_iter.first;
-
-        if (!result_iter.second) {
-            if (debug) {
-                std::cout << region_idx << ": Seed analysis failed at seed " << seed_pos.x << ", " << seed_pos.y << "=> iterative search failed" << std::endl;
-            }
-            result.success = false;
-            continue;
-        }
-        
-        // Extract optimized parameters from result
-        SeedParams seed_result;
-        seed_result.x = seed_pos.x;
-        seed_result.y = seed_pos.y;
-        seed_result.v = params_result(2);
-        seed_result.u = params_result(3);
-        seed_result.dv_dx = params_result(4);
-        seed_result.dv_dy = params_result(5);
-        seed_result.du_dx = params_result(6);
-        seed_result.du_dy = params_result(7);
-        seed_result.corrcoef = params_result(8);
-        
-        // Extract convergence metrics
-        SeedConvergence convergence;
-        convergence.num_iterations = sr_nloptimizer.get_last_iteration_count();  // Actual count
-        convergence.diffnorm = params_result(9);
-        
-        result.seeds.push_back(seed_result);
-        result.convergence.push_back(convergence);
-        
-        // Check quality thresholds
-        if (seed_result.corrcoef > cutoff_max_corrcoef || convergence.diffnorm > cutoff_max_diffnorm) {
-            if (debug) {
-                std::cout << region_idx << ": Seed analysis failed at seed " << seed_pos.x << ", " << seed_pos.y << "=> corrcoef: " << seed_result.corrcoef << ", diffnorm: " << convergence.diffnorm << std::endl;
-            }
-            std::cout << result.success << std::endl;
-            result.success = false; //TODO: Solve this
-        }
-        region_idx++;
-    }
-    
-    return result;
-}
-
-namespace {
-
-using matlab_difference_type = ROI2D::difference_type;
-
-struct matlab_seed_segment final {
-    matlab_difference_type ref_idx = 0;
-    std::vector<std::vector<SeedParams>> seeds_by_frame;
-    std::vector<SeedParams> terminal_seeds;
-};
-
-Array2D<double> matlab_make_seed_params_init(const SeedParams& seed) {
-    Array2D<double> params_init(10, 1);
-    params_init(0) = seed.y;
-    params_init(1) = seed.x;
-    params_init(2) = 0.0;
-    params_init(3) = 0.0;
-    params_init(4) = 0.0;
-    params_init(5) = 0.0;
-    params_init(6) = 0.0;
-    params_init(7) = 0.0;
-    params_init(8) = 0.0;
-    params_init(9) = 0.0;
-    return params_init;
-}
-
-bool matlab_seed_matches_region(const ROI2D& roi_reduced,
-                                const SeedParams& seed,
-                                matlab_difference_type scalefactor,
-                                matlab_difference_type region_idx) {
-    if (scalefactor <= 0) {
-        return false;
-    }
-    const matlab_difference_type reduced_y = seed.y / scalefactor;
-    const matlab_difference_type reduced_x = seed.x / scalefactor;
-    const auto region_idx_pair = roi_reduced.get_region_idx(reduced_y, reduced_x);
-    return region_idx_pair.first == region_idx;
-}
-
-bool matlab_seed_positions_are_unique(const std::vector<SeedParams>& seeds,
-                                      matlab_difference_type scalefactor) {
-    for (std::size_t i = 0; i < seeds.size(); ++i) {
-        for (std::size_t j = i + 1; j < seeds.size(); ++j) {
-            if (seeds[i].x / scalefactor == seeds[j].x / scalefactor &&
-                seeds[i].y / scalefactor == seeds[j].y / scalefactor) {
-                return false;
-            }
-        }
-    }
-    return true;
-}
-
-std::vector<SeedParams> matlab_propagate_seeds(const std::vector<SeedParams>& seeds,
-                                               matlab_difference_type scalefactor) {
-    std::vector<SeedParams> propagated_seeds;
-    propagated_seeds.reserve(seeds.size());
-
-    for (const auto& seed : seeds) {
-        SeedParams propagated = seed;
-        propagated.x = seed.x + static_cast<matlab_difference_type>(std::round(seed.u / scalefactor) * scalefactor);
-        propagated.y = seed.y + static_cast<matlab_difference_type>(std::round(seed.v / scalefactor) * scalefactor);
-        propagated.u = 0.0;
-        propagated.v = 0.0;
-        propagated.du_dx = 0.0;
-        propagated.du_dy = 0.0;
-        propagated.dv_dx = 0.0;
-        propagated.dv_dy = 0.0;
-        propagated.corrcoef = 0.0;
-        propagated_seeds.push_back(propagated);
-    }
-
-    return propagated_seeds;
-}
-
-matlab_seed_segment matlab_compute_seed_segment(const DIC_analysis_parallel_input& input,
-                                                matlab_difference_type ref_idx,
-                                                const ROI2D& roi_ref,
-                                                const std::vector<SeedParams>& seeds_by_region,
-                                                bool seeds_are_optimized) {
-    typedef ROI2D::difference_type difference_type;
-
-    const auto& DIC_input = input.base_input;
-    matlab_seed_segment segment;
-    segment.ref_idx = ref_idx;
-
-    if (seeds_by_region.empty()) {
-        return segment;
-    }
-
-    auto roi_reduced = roi_ref.reduce(DIC_input.scalefactor);
-    if (difference_type(seeds_by_region.size()) != roi_reduced.size_regions()) {
-        throw std::invalid_argument("matlab_DIC_analysis requires seeds_by_region.size() to match the number of ROI regions.");
-    }
-
-    std::vector<SeedParams> predicted_seeds = seeds_are_optimized ?
-        matlab_propagate_seeds(seeds_by_region, DIC_input.scalefactor) :
-        seeds_by_region;
-
-    if (!matlab_seed_positions_are_unique(predicted_seeds, DIC_input.scalefactor)) {
-        throw std::invalid_argument("matlab_DIC_analysis received duplicate seed positions on the reduced grid.");
-    }
-
-    segment.seeds_by_frame.reserve(DIC_input.imgs.size() - ref_idx - 1);
-
-    const auto A_ref = DIC_input.imgs[ref_idx].get_gs();
-    for (difference_type cur_idx = ref_idx + 1; cur_idx < difference_type(DIC_input.imgs.size()); ++cur_idx) {
-        const auto A_cur = DIC_input.imgs[cur_idx].get_gs();
-        auto sr_nloptimizer = details::subregion_nloptimizer(
-            A_ref,
-            A_cur,
-            roi_ref,
-            DIC_input.scalefactor,
-            DIC_input.interp_type,
-            DIC_input.subregion_type,
-            DIC_input.r
-        );
-
-        std::vector<SeedParams> optimized_seeds;
-        optimized_seeds.reserve(predicted_seeds.size());
-        bool frame_success = matlab_seed_positions_are_unique(predicted_seeds, DIC_input.scalefactor);
-
-        for (difference_type region_idx = 0;
-             frame_success && region_idx < difference_type(predicted_seeds.size());
-             ++region_idx) {
-            const auto& predicted_seed = predicted_seeds[region_idx];
-            if (!matlab_seed_matches_region(roi_reduced, predicted_seed, DIC_input.scalefactor, region_idx)) {
-                if (DIC_input.debug) {
-                    std::cout << "matlab_DIC_analysis: seed for region " << region_idx
-                              << " is outside its ROI at frame " << cur_idx << "." << std::endl;
-                }
-                frame_success = false;
-                break;
-            }
-
-            auto params_init = matlab_make_seed_params_init(predicted_seed);
-            auto global_result = sr_nloptimizer.global(params_init);
-            if (!global_result.second) {
-                if (DIC_input.debug) {
-                    std::cout << "matlab_DIC_analysis: global seed optimization failed for region "
-                              << region_idx << " at frame " << cur_idx << "." << std::endl;
-                }
-                frame_success = false;
-                break;
-            }
-
-            SeedParams optimized_seed = SeedParams::from_array(global_result.first);
-            const double diffnorm = global_result.first(9);
-            const int num_iterations = sr_nloptimizer.get_last_iteration_count();
-            
-            // MATLAB-style checks: corrcoef, diffnorm, and iteration saturation
-            // MATLAB ncorr_abr_alg_seedanalysis.m line 85: (i > 0 && any([convergence_buffer.num_iterations] == cutoff_iteration))
-            bool iteration_saturated = (cur_idx > ref_idx + 1) && (num_iterations >= 100);  // Default cutoff_iterations = 100
-            
-            if (optimized_seed.corrcoef > input.cutoff_max_corrcoef ||
-                diffnorm > input.cutoff_max_diffnorm ||
-                iteration_saturated ||
-                !matlab_seed_matches_region(roi_reduced, optimized_seed, DIC_input.scalefactor, region_idx)) {
-                if (DIC_input.debug) {
-                    std::cout << "matlab_DIC_analysis: seed quality failed for region " << region_idx
-                              << " at frame " << cur_idx
-                              << " corrcoef=" << optimized_seed.corrcoef
-                              << " diffnorm=" << diffnorm
-                              << " iterations=" << num_iterations
-                              << " saturated=" << (iteration_saturated ? "yes" : "no") << "." << std::endl;
-                }
-                frame_success = false;
-                break;
-            }
-
-            optimized_seeds.push_back(optimized_seed);
-        }
-
-        if (!frame_success || !matlab_seed_positions_are_unique(optimized_seeds, DIC_input.scalefactor)) {
-            if (DIC_input.debug) {
-                std::cout << "matlab_DIC_analysis: stopping seed schedule at frame " << cur_idx << "." << std::endl;
-            }
-            break;
-        }
-
-        segment.seeds_by_frame.push_back(optimized_seeds);
-        // NOTE: Do NOT propagate seeds within a segment.
-        // MATLAB uses the SAME fixed pos_seed for ALL frames in a segment
-        // (ncorr_abr_alg_seedanalysis.m calls ncorr_alg_calcseeds with the
-        // same pos_seed for each current image). Propagation only happens
-        // BETWEEN segments in matlab_DIC_analysis_impl.
-    }
-
-    if (!segment.seeds_by_frame.empty()) {
-        segment.terminal_seeds = segment.seeds_by_frame.back();
-    }
-
-    return segment;
-}
-
-std::vector<Disp2D> matlab_run_segment_dic(const DIC_analysis_parallel_input& input,
-                                           const ROI2D& roi_ref,
-                                           const matlab_seed_segment& segment,
-                                           bool run_in_parallel) {
-    typedef ROI2D::difference_type difference_type;
-
-    std::vector<Disp2D> segment_disps(segment.seeds_by_frame.size());
-    if (segment.seeds_by_frame.empty()) {
-        return segment_disps;
-    }
-
-    const auto& DIC_input = input.base_input;
-    const auto A_ref = DIC_input.imgs[segment.ref_idx].get_gs();
-    const difference_type num_frames = static_cast<difference_type>(segment.seeds_by_frame.size());
-
-    auto compute_frame = [&](difference_type frame_idx) {
-        const difference_type cur_idx = segment.ref_idx + frame_idx + 1;
-        const auto A_cur = DIC_input.imgs[cur_idx].get_gs();
-        return RGDIC_without_thread(
-            A_ref,
-            A_cur,
-            roi_ref,
-            DIC_input.scalefactor,
-            DIC_input.interp_type,
-            DIC_input.subregion_type,
-            DIC_input.r,
-            DIC_input.cutoff_corrcoef,
-            DIC_input.debug,
-            segment.seeds_by_frame[frame_idx],
-            true
-        );
-    };
-
-    if (!run_in_parallel || num_frames <= 1) {
-        for (difference_type frame_idx = 0; frame_idx < num_frames; ++frame_idx) {
-            segment_disps[frame_idx] = compute_frame(frame_idx);
-        }
-        return segment_disps;
-    }
-
-    #pragma omp parallel for num_threads(std::min(num_frames, DIC_input.num_threads)) schedule(dynamic)
-    for (difference_type frame_idx = 0; frame_idx < num_frames; ++frame_idx) {
-        segment_disps[frame_idx] = compute_frame(frame_idx);
-    }
-
-    return segment_disps;
-}
-
-DIC_analysis_output matlab_DIC_analysis_impl(const DIC_analysis_parallel_input& input,
-                                             bool run_in_parallel,
-                                             const std::string& mode_name) {
-    typedef ROI2D::difference_type difference_type;
-
-    const auto& DIC_input = input.base_input;
-    if (DIC_input.imgs.size() < 2) {
-        throw std::invalid_argument("matlab_DIC_analysis requires at least 2 images.");
-    }
-    if (input.seeds_by_region.empty()) {
-        throw std::invalid_argument("matlab_DIC_analysis requires one manual/preset seed per ROI region.");
-    }
-
-    DIC_analysis_output DIC_output;
-    DIC_output.disps.resize(DIC_input.imgs.size() - 1);
-    DIC_output.perspective_type = PERSPECTIVE::LAGRANGIAN;
-    DIC_output.units = "pixels";
-    DIC_output.units_per_pixel = 1.0;
-
-    std::vector<Disp2D> step_disps;
-    std::vector<ROI2D> step_rois;
-    std::vector<difference_type> step_ref_idx;
-    if (DIC_input.save_disps_steps) {
-        step_disps.resize(DIC_output.disps.size());
-        step_rois.resize(DIC_output.disps.size());
-        step_ref_idx.resize(DIC_output.disps.size());
-    }
-
-    difference_type ref_idx = 0;
-    ROI2D roi_ref = DIC_input.roi;
-    std::vector<SeedParams> current_seeds = input.seeds_by_region;
-    bool current_seeds_optimized = input.seeds_are_optimized;
-
-    while (ref_idx < difference_type(DIC_input.imgs.size()) - 1) {
-        const auto segment = matlab_compute_seed_segment(
-            input,
-            ref_idx,
-            roi_ref,
-            current_seeds,
-            current_seeds_optimized
-        );
-
-        if (segment.seeds_by_frame.empty()) {
-            if (ref_idx == 0) {
-                throw std::runtime_error(mode_name + " could not seed any current image.");
-            }
-            throw std::runtime_error(
-                mode_name + " could not seed the segment starting at reference frame " +
-                std::to_string(ref_idx + 1) + "."
-            );
-        }
-
-        const auto segment_disps = matlab_run_segment_dic(input, roi_ref, segment, run_in_parallel);
-        ROI2D next_roi_ref;
-        for (difference_type frame_idx = 0; frame_idx < difference_type(segment_disps.size()); ++frame_idx) {
-            const difference_type cur_idx = ref_idx + frame_idx + 1;
-            DIC_output.disps[cur_idx - 1] = segment_disps[frame_idx];
-            next_roi_ref = matlab_update_roi(roi_ref, segment_disps[frame_idx], DIC_input.interp_type, DIC_input.r);
-
-            if (DIC_input.save_disps_steps) {
-                step_disps[cur_idx - 1] = segment_disps[frame_idx];
-                step_rois[cur_idx - 1] = roi_ref;
-                step_ref_idx[cur_idx - 1] = ref_idx;
-            }
-        }
-
-        const difference_type segment_end_idx = ref_idx + difference_type(segment.seeds_by_frame.size());
-        if (DIC_input.debug) {
-            std::cout << "matlab_DIC_analysis: processed segment "
-                      << (ref_idx + 1) << " -> " << (segment_end_idx + 1)
-                      << " (" << segment.seeds_by_frame.size() << " frame(s))." << std::endl;
-        }
-
-        ref_idx = segment_end_idx;
-        if (ref_idx >= difference_type(DIC_input.imgs.size()) - 1) {
-            break;
-        }
-
-        roi_ref = next_roi_ref;
-        current_seeds = segment.terminal_seeds;
-        current_seeds_optimized = true;
-    }
-
-    if (DIC_input.save_disps_steps && !step_disps.empty()) {
-        DIC_analysis_step_data step_data;
-        step_data.step_disps = step_disps;
-        step_data.step_rois = step_rois;
-        step_data.step_ref_idx = step_ref_idx;
-
-        const std::string step_filename = run_in_parallel ?
-            "matlab_DIC_analysis_parallel_step_data.bin" :
-            "matlab_DIC_analysis_sequential_step_data.bin";
-        save(step_data, step_filename);
-        std::cout << "Step displacement data saved to " << step_filename << std::endl;
-    }
-
-    return DIC_output;
-}
-
-} // namespace
-
-// Matlab-ABR-style DIC analysis -------------------------------------------//
-DIC_analysis_output matlab_DIC_analysis_sequential(const DIC_analysis_input &DIC_input,
-                                                   const std::vector<SeedParams>& seeds_by_region,
-                                                   bool seeds_are_optimized) {
-    return matlab_DIC_analysis_sequential(DIC_analysis_parallel_input(DIC_input, seeds_by_region, seeds_are_optimized));
-}
-
-DIC_analysis_output matlab_DIC_analysis_sequential(const DIC_analysis_parallel_input &input) {
-    return matlab_DIC_analysis_impl(input, false, "matlab_DIC_analysis_sequential");
-}
-
-DIC_analysis_output matlab_DIC_analysis_parallel(const DIC_analysis_parallel_input& input) {
-    return matlab_DIC_analysis_impl(input, true, "matlab_DIC_analysis_parallel");
-}
-
-// Parallel DIC analysis using seed-based failure prediction
-DIC_analysis_output DIC_analysis_parallel(const DIC_analysis_parallel_input& input) {
-    typedef std::ptrdiff_t difference_type;
-    
-    const auto& DIC_input = input.base_input;
-    
-    if (DIC_input.imgs.size() < 2) {
-        throw std::invalid_argument("DIC_analysis_parallel requires at least 2 images.");
-    }
-    
-    // Seeds should be provided per region (not generated here)
-    if (input.seeds_by_region.empty()) {
-        throw std::invalid_argument("DIC_analysis_parallel requires seeds_by_region to be provided.");
-    }
-    
-    std::cout << std::endl << "Starting seed-based parallel DIC analysis..." << std::endl;
-    std::cout << "Number of regions: " << input.seeds_by_region.size() << std::endl;
-    std::cout << "Cutoff max diffnorm: " << input.cutoff_max_diffnorm << std::endl;
-    std::cout << "Cutoff max corrcoef: " << input.cutoff_max_corrcoef << std::endl;
-    
-    std::chrono::time_point<std::chrono::system_clock> start_analysis = std::chrono::system_clock::now();
-    
-    // Initialize output
-    DIC_analysis_output DIC_output;
-    DIC_output.disps.resize(DIC_input.imgs.size() - 1);
-    DIC_output.perspective_type = PERSPECTIVE::LAGRANGIAN;
-    DIC_output.units = "pixels";
-    DIC_output.units_per_pixel = 1.0;
-    
-    // PHASE 1: Compute seed parameters for all frames with ROI updates
-    std::cout << "\nPhase 1: Computing seed parameters for all frames..." << std::endl;
-    
-    // Prepare current images
-    std::vector<Array2D<double>> A_curs;
-    A_curs.reserve(DIC_input.imgs.size() - 1);
-    for (size_t i = 1; i < DIC_input.imgs.size(); ++i) {
-        A_curs.push_back(DIC_input.imgs[i].get_gs());
-    }
-    
-    // Compute seed parameters upfront
-    std::vector<SeedComputationData> seed_data = compute_only_seed_points(
-        DIC_input.imgs[0].get_gs(),
-        A_curs,
-        DIC_input.roi,
-        DIC_input.scalefactor,
-        DIC_input.interp_type,
-        DIC_input.subregion_type,
-        DIC_input.r,
-        input.seeds_by_region,
-        DIC_input.cutoff_corrcoef,
-        0,  // region_idx (working with first region, but structure supports n regions)
-        DIC_input.debug
-    );
-    
-    if (seed_data.empty()) {
-        throw std::runtime_error("No valid seed parameters could be computed.");
-    }
-    
-    std::cout << "Successfully computed seed parameters for " << seed_data.size() << " frames" << std::endl;
-    
-    // PHASE 2: Parallel displacement computation using precomputed seed data
-    std::cout << "\nPhase 2: Computing displacements in parallel..." << std::endl;
-    
-    size_t safe_batch_size = seed_data.size();
-    
-    // Use OpenMP for parallel execution (Python uses ThreadPoolExecutor, C++ uses OpenMP)
-    std::cout << "Safe batch size: " << safe_batch_size << std::endl;
-    std::cout << "Number of threads: " << DIC_input.num_threads << std::endl;
-    std::cout << "static_cast<difference_type>(safe_batch_size): " << static_cast<difference_type>(safe_batch_size) << std::endl;
-    std::cout << "Number of threads: " << std::min(static_cast<difference_type>(safe_batch_size), DIC_input.num_threads) << std::endl;
-    #pragma omp parallel for num_threads(std::min(static_cast<difference_type>(safe_batch_size), DIC_input.num_threads)) schedule(dynamic)
-    for (size_t i = 0; i < safe_batch_size; ++i) {
-        difference_type frame_idx = static_cast<difference_type>(i) + 1;
-        int thread_id = omp_get_thread_num();
-        int total_threads = omp_get_num_threads();
-        printf("Running on thread %d of %d\n", thread_id, total_threads);
-        
-        #pragma omp critical
-        {
-            std::cout << "  Processing frame " << frame_idx << " (parallel)" << std::endl;
-        }
-        
-        // Use precomputed seed data for this frame
-        const auto& frame_data = seed_data[i];
-        const auto& roi_for_frame = frame_data.roi;
-        const auto& sr_nloptimizer_for_frame = frame_data.sr_nloptimizer;
-        
-        // Safety check: ensure seed_params_by_region is not empty
-        if (frame_data.seed_params_by_region.empty()) {
-            #pragma omp critical
-            {
-                std::cerr << "  WARNING: No seed params for frame " << frame_idx << ", skipping." << std::endl;
-            }
-            continue;
-        }
-        const auto& seed_params_for_frame = frame_data.seed_params_by_region[0];
-        
-        try {
-            // Compute displacements using precomputed seed parameters
-            auto disps = compute_displacements(
-                sr_nloptimizer_for_frame,
-                roi_for_frame.reduce(DIC_input.scalefactor),
-                seed_params_for_frame,
-                DIC_input.scalefactor,
-                DIC_input.cutoff_corrcoef,
-                0,  // region_idx
-                DIC_input.debug
-            );
-            
-            // Store displacement
-            #pragma omp critical
-            {
-                DIC_output.disps[frame_idx - 1] = disps;
-            }
-        } catch (const std::exception& e) {
-            #pragma omp critical
-            {
-                std::cerr << "  ERROR: Frame " << frame_idx << " failed: " << e.what() << std::endl;
-            }
-        }
-    }
-    
-    // End timer
-    std::chrono::time_point<std::chrono::system_clock> end_analysis = std::chrono::system_clock::now();
-    std::chrono::duration<double> elapsed_seconds = end_analysis - start_analysis;
-    std::cout << std::endl << "Total parallel DIC analysis time: " << elapsed_seconds.count() << " seconds" << std::endl;
-    
-    return DIC_output;
-}
-
-}
+} // namespace ncorr
